Add file identifier and offset support in Line
Line objects recieved from the tailer now include a file identifier and
offset as some of the context. This limits the cost of making syscalls
to seek in the file to find out the current position. The file
identifier will be used to only seek within the same individual file if
provided.