Only seek if the file identifier matches
When tailing files that are rotated, we do not want to seek into a file
at the last offset if the file has been rotated. We use the
fileIdentifier we track when opening files (based on inode and device on
posix) to skip the seek if the property is set and the current file
doesn't match.