Update inotify to use os.Stat exclusively for creations
The previous mechanism of watching parent directories using inotify is
problematic when the parent directories do not yet exist or when
symlinks are pointing at the actual files being watched. Symlinks are
very common in the Kubernetes space where log files are pointed to with
longer chains of symlinks.

os.Stat and the inotify calls made by fsnotify/fsnotify follow symlinks
down and correctly record the status of the underlying files. For
consistency, when we wait for the creation of a file we wish to watch we
need to use a consistent mechanism. Since we cannot use inotify with
files that do not exist, instead just poll with os.Stat until it is
created.