Update watch implementations for less racieness
The prior API was racy or could lead to incorrect results when watching
for changes in files. It depended on recreating the watches fully when
it detected a file rotated and was stateful with its own loops, but
could easily get out of sync from what files we actually had open in the
tailer. It also did not handle errors from inotify correctly, meaning
that it would not recreate the fd when necessary (e.g. during overflow).

We now also handle cases where a file is rotated, but we have not
consumed all of the data from the prior open file. If we detect
rotation, we attempt one more set of reads to read whatever contents
were added to the file between our last EOF and the flush done by any
writer prior to rotation. This allows us to capture the ends of files
consistently.