Change FileIdentifier interface to take *os.File
This allows the library on Windows to more-correctly extract a
reasonable opaque file identifier string. The fileInfo struct populated
by the go stdlib does include the necessary idxHi, idxLo, and volume
serial ids but they are all private. os.FileInfo.Sys() on Windows
returns an object that contains file modification dates and the like,
but not the info we need for this string creation.

Instead, we pass the *os.File instance in. On posix, just call os.Stat
to get the inode and device. On Windows, call
GetFileInformationByHandle.